Full project
The whole project. Controls are now adjusted for mobile devices, loading screen, timer, small UI, camera zoom, effects, dance at the end is randomly selected, a lot of things completed...
And I like the project! Maybe in the future there will be some more changes, and this is most likely in the code, because there is a lot to streamline there...